+++ packages/provider/lib/provider.dart
@@ -3,7 +3,7 @@ library provider;
 export 'src/async_provider.dart';
 export 'src/change_notifier_provider.dart';
 export 'src/consumer.dart';
-export 'src/inherited_provider.dart' hide autoDeferred;
+export 'src/inherited_provider.dart' hide autoDeferred, isWidgetTreeBuilding;
 export 'src/listenable_provider.dart';
 export 'src/provider.dart';
 export 'src/proxy_provider.dart'